--- crates/ibc/src/bigint.rs
@@ -1,1 +1,0 @@
-pub use primitive_types::U256;
