+++ crates/ibc/src/core/ics04_channel/handler/recv_packet.rs
@@ -180,7 +180,7 @@ mod tests {
     use crate::core::ics04_channel::Version;
     use crate::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
     use crate::mock::context::MockContext;
-    use crate::relayer::ics18_relayer::context::RelayerContext;
+    use crate::mock::ics18_relayer::context::RelayerContext;
     use crate::test_utils::get_dummy_account_id;
     use crate::timestamp::Timestamp;
     use crate::timestamp::ZERO_DURATION;
