--- crates/ibc/src/relayer/mod.rs
@@ -1,5 +1,0 @@
-//! Utilities for testing the `ibc` crate against the [Hermes IBC relayer][relayer-repo].
-//!
-//! [relayer-repo]: https://github.com/informalsystems/ibc-rs/tree/main/relayer
-
-pub mod ics18_relayer;
