+++ crates/ibc/src/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod macros;
 pub mod pretty;
