+++ crates/ibc/src/utils/pretty.rs
@@ -1,17 +1,8 @@
-use core::fmt::{Debug, Display, Error as FmtError, Formatter};
-use core::time::Duration;
+use core::fmt::{Display, Error as FmtError, Formatter};
 use tendermint::block::signed_header::SignedHeader;
 use tendermint::validator::Set as ValidatorSet;
 
 use alloc::vec::Vec;
-
-pub struct PrettyDuration<'a>(pub &'a Duration);
-
-impl Display for PrettyDuration<'_> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), FmtError> {
-        Debug::fmt(self.0, f)
-    }
-}
 
 pub struct PrettyOption<'a, T>(pub &'a Option<T>);
 
@@ -85,36 +76,6 @@ mod tests {
 
     use crate::alloc::string::ToString;
     use std::{string::String, vec};
-
-    #[test]
-    fn test_pretty_duration_micros() {
-        let expected_output = "5µs";
-
-        let duration = Duration::from_micros(5);
-        let pretty_duration = PrettyDuration(&duration);
-
-        assert_eq!(pretty_duration.to_string(), expected_output);
-    }
-
-    #[test]
-    fn test_pretty_duration_millis() {
-        let expected_output = "5ms";
-
-        let duration = Duration::from_millis(5);
-        let pretty_duration = PrettyDuration(&duration);
-
-        assert_eq!(pretty_duration.to_string(), expected_output);
-    }
-
-    #[test]
-    fn test_pretty_duration_secs() {
-        let expected_output = "10s";
-
-        let duration = Duration::from_secs(10);
-        let pretty_duration = PrettyDuration(&duration);
-
-        assert_eq!(pretty_duration.to_string(), expected_output);
-    }
 
     #[test]
     fn test_pretty_option_some() {
